pyproject.toml
[tool.poetry]
name = "javad-amin-github-io"
version = "0.1.0"
description = ""
authors = ["javad-amin"]
readme = "README.md"
packages = [{ include = "javad_amin" }]

[tool.poetry.dependencies]
python = "^3.12"
pyyaml = "^6.0.2"
jinja2 = "^3.1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
